add logic to get jira issue search results by project and convert to a simple class

+++ src/jr.common/Jira/Models/JiraIssueResults.cs
@@ -8,12 +8,12 @@ namespace jr.common.Jira.Models
         public JiraIssue[] Issues { get; set; }
 
         [JsonProperty("maxResults")]
-        public long MaxResults { get; set; }
+        public int MaxResults { get; set; }
 
         [JsonProperty("startAt")]
-        public long StartAt { get; set; }
+        public int StartAt { get; set; }
 
         [JsonProperty("total")]
-        public long Total { get; set; }
+        public int Total { get; set; }
     }
 }
